feat: Send OOR payment from Alice to Bob
The fix involved correctly encoding the relative locktime in the
Miniscript. We have to use a special consensus encoding. Maybe there
is a Miniscript API that already offers this for us, but I could not
find it.

Relatedly, I fixed a TODO where we were hard-coding the timeout of the
BOARDING_DESCRIPTOR_TEMPLATE_MINISCRIPT.