feat: Implement Ark transaction history
The implementation is directly taken from the Go client. The tests are
also taken from the Go client.

I believe there are a few bugs in the original implementation, but for
now we aim for feature parity. Another focus was for the
implementation to be clearer and more readable (YMMV) and to include
more comments.

Additionally, we implement a transaction history command in
`ark-sample`.

Co-authored-by: Philipp Hoenisch <philipp@coblox.tech>