fix: Only submit round PSBT if we have boarding inputs
Even when we did not have boarding inputs for a roun, we were calling
`sign_round_psbt`. This would result in no changes to the `round_psbt`
received in the event. We would then submit an unchanged round
PSBT (together with some signed forfeit PSBTs) and the server would
sometimes trip on this due to a race condition.