fix: allow to parse different network from ark server
Ark server returns `mutinynet` which is not a `bitcoin::Network` type. To not have to handle internally with our custom network type we introduce a `Network` type  only when parsing responses from Ark and convert directly to `bitcoin::Network`. Note: right now `info` is the only response which includes a `Network`, if this type is needed elsewhere we should move it.