feat: Use rust-secp256k1 branch for Musig
We were using a `rust-secp256k1-zkp` branch for Musig, but it's better
to use the branch that will eventually merge into `rust-secp256k1`.

Incidentally, we still need the dependency on `rust-secp256k1-zkp` in
`ark-dlc-sample`, because we are using Musig adaptor signatures for
that one. But that can be a regular git dependency, since
`ark-dlc-sample` will not be released to crates.io.